controllers/organizations/sessions_controller: Remove unused configure_sign_in_params() method.
This was added in b7c5916e6. As far as I can tell, this was
misguided. The intended pattern is to also uncomment the top call to
before_action(), whereupon configure_sign_in_params() would be called
as a hook. However, that was not done. Since we do the actual
authentication inside of the Organizations model, this was unnecessary
anyway.

Beyond that it's also not clear to me how this was supposed to have
worked: as I understand (and the documentation is as always not quite
clear), Devise parameter sanitation is for parameters passed in
through the POST request. Therefore this code would have had the
effect of allowing an "approved" parameter to be posted by the user,
which is several steps away from what we actually want to have happen.